Fixed memory leaks. Had to do some complicated stuff to get bound types to be registered so they actually get cleaned up but will still get auto created if they didn't exist.

+++ Reflection/TypeId.hpp
@@ -3,6 +3,6 @@
 struct TypeId
 {
   TypeId() : mId(0) {}
-  TypeId(int id) : mId(id) {}
+  TypeId(size_t id) : mId(id) {}
   size_t mId;
 };
